refactor: Move creation of baml dynamic type to own file
